Merck/requirements.txt
# Core AutoGen and AI
autogen>=0.2.0
openai>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
xlrd>=2.0.1

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# Document Processing (for multimodal LLM)
PyPDF2>=3.0.0
Pillow>=10.0.0
python-docx>=0.8.11
pdf2image>=1.16.0

# Web and API
requests>=2.31.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Email and Notifications
email-validator>=2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
loguru>=0.7.0
python-dateutil>=2.8.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0 
